Adjust display brightness from sensor reading.

+++ include/ClockDisplay.h
@@ -4,7 +4,7 @@
 namespace ClockDisplay {
     void begin(void);
     void displayTime(uint8_t hour, uint8_t minute);
-    void adjustBrightness(short brightness);
+    void adjustBrightness(long brightness);
     void switchDisplay(bool state);
     void displayError(int errorCode);
 }
